requirements.txt
python-dotenv
requests
msal
pandas
openpyxl
py-zabbix
